requirements-dev.txt
-r requirements.txt
pytest
requests-mock
paho-mqtt
